cdk_bootstrap/requirements.txt
# cdk stuff
aws-cdk-lib
constructs
